chore: modernize ruff configuration
Use human readable names
